# docs/openapi/client.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Eau La Maman API",
        "version": "1.0.0",
        "description": "API consomm\u00e9e par l'espace client Nuxt et l'application mobile (livreurs/propri\u00e9taires). Authentification par token Bearer Sanctum \u2014 voir docs/api-auth-contract.md et docs/api-espace-client-contract.md pour les d\u00e9cisions m\u00e9tier et les limites connues que cette sp\u00e9cification ne peut pas exprimer."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local"
        },
        {
            "url": "https://formation.eau-la-maman.com/api",
            "description": "Pr\u00e9prod"
        },
        {
            "url": "https://fello.eau-la-maman.com/api",
            "description": "Production"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/auth/otp-login/request": {
            "post": {
                "operationId": "auth.otp-login.request",
                "tags": [
                    "Autres"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "telephone": {
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "telephone"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "sent": {
                                            "type": "boolean"
                                        },
                                        "channel": {
                                            "type": "string"
                                        },
                                        "destination_masked": {
                                            "type": "string",
                                            "description": "Ajout\u00e9 le 27/08/2026 (demande front) : la coordonn\u00e9e R\u00c9ELLEMENT\nutilis\u00e9e pour ce canal, d\u00e9j\u00e0 masqu\u00e9e c\u00f4t\u00e9 serveur (cf.\nOtpDestinationMasker) \u2014 jamais l'adresse/le num\u00e9ro complet,\njamais reconstruite c\u00f4t\u00e9 client \u00e0 partir d'une autre source."
                                        },
                                        "cooldown_seconds": {
                                            "type": "integer",
                                            "const": 30
                                        }
                                    },
                                    "required": [
                                        "sent",
                                        "channel",
                                        "destination_masked",
                                        "cooldown_seconds"
                                    ]
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "anyOf": [
                                        {
                                            "type": "object",
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "const": "Impossible d'envoyer un code de connexion \u00e0 ce compte pour le moment."
                                                }
                                            },
                                            "required": [
                                                "error"
                                            ]
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "error": {
                                                    "type": "string",
                                                    "const": "Aucun canal disponible pour recevoir un code de connexion pour le moment."
                                                }
                                            },
                                            "required": [
                                                "error"
                                            ]
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        },
                                        "retry_after_seconds": {
                                            "type": "integer"
                                        }
                                    },
                                    "required": [
                                        "error",
                                        "retry_after_seconds"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "const": "Aucun compte trouv\u00e9 pour ce num\u00e9ro de t\u00e9l\u00e9phone."
                                        }
                                    },
                                    "required": [
                                        "error"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                },
                "security": []
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "description": "Token Sanctum obtenu via `POST /api/auth/login` (champ `token` de la r\u00e9ponse). Envoyer `Authorization: Bearer <token>` sur toute route prot\u00e9g\u00e9e. Coller un token de d\u00e9veloppement ici (bouton Authorize) pour tester les endpoints prot\u00e9g\u00e9s \u2014 jamais un token d'un compte r\u00e9el en production.",
                "scheme": "bearer",
                "bearerFormat": "Sanctum"
            }
        },
...
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
...
        }
    }
}